+++ src/main/java/net/rewardz/RewardzMain.java
@@ -6,6 +6,7 @@ import java.util.Map;
 
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.loader.api.FabricLoader;
+import net.rewardz.init.CommandInit;
 import net.rewardz.init.ConfigInit;
 import net.rewardz.init.LoaderInit;
 import net.rewardz.init.ScreenInit;
@@ -14,7 +15,6 @@ import net.rewardz.packet.RewardsServerPacket;
 public class RewardzMain implements ModInitializer {
 
     public static final boolean isPatchouliButtonLoaded = FabricLoader.getInstance().isModLoaded("patchoulibutton");
-    // public static final Map<Integer, List<Object>> REWARD_MAP = new HashMap<Integer, List<Object>>();
     public static final Map<Integer, Map<Integer, List<Object>>> REWARD_MAP = new HashMap<Integer, Map<Integer, List<Object>>>();
 
     @Override
@@ -22,6 +22,7 @@ public class RewardzMain implements ModInitializer {
         ConfigInit.init();
         LoaderInit.init();
         ScreenInit.init();
+        CommandInit.init();
         RewardsServerPacket.init();
     }